pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "designpilot_fusion"
version = "0.1.0"
description = "DesignPilot add-in for Fusion."
readme = "README.md"
requires-python = ">=3.12" # Development version
dependencies = [
    # "rxprop>=0.1.0",
    # Development versions
    "case-converter>=1.2.0",
    "jinja2>=3.1.6",
    "litellm>=1.72.0",
    "markdown-it-py>=3.0.0"
]
classifiers = []


[tool.setuptools]
package-dir = {"" = "Source/Packages"}


[tool.pytest.ini_options]
pythonpath = ["Source"]
timeout = 3
testpaths = [ "Tests/", "External/rxprop/Tests/" ]


[tool.pyright]
typeCheckingMode = "strict"
executionEnvironments = [
    { root = "Tests", typeCheckingMode = "basic" }
]
reportMissingTypeStubs = "none"
stubPath = "Source/Stubs"
